Fa02 create unigrams (#8)
* Establish Gradle Project

This commit makes a simple gradle project with Mapper, Reducer, and Job files. The code inside the files for now will run the word count example. This is so you can compile the code and create a jar.

* Update README

Add commands to compile and package code

* Create Unigrams with Document ID

This PR extracts the document ID and each word in the file and creates intermediate key-value pairs